remove unused provider in tx
