+++ src/components/ActiveTapplet.tsx
@@ -10,20 +10,7 @@ import { ErrorSource } from "../store/error/error.types"
 import { LaunchedTappResult } from "@type/tapplet"
 import { tappletProvidersActions } from "../store/tappletProviders/tappletProviders.slice"
 import { RootState } from "../store/store"
-import { tappletProviderSelector } from "../store/tappletProviders/tappletProviders.selector"
-
-const selectTappProviderById = (state: RootState, id?: string) =>
-  id ? tappletProviderSelector.getTappletProviderById(state, id) : null
-
-export type TappProviderIdParam = {
-  installedTappletId?: number | string
-  devTappletId?: number | string
-}
-export const getTappProviderId = (tappId: TappProviderIdParam): string => {
-  if (tappId.devTappletId) return `DTP${tappId.devTappletId}`
-  if (tappId.installedTappletId) return `TP${tappId.installedTappletId}`
-  return "undefined-id"
-}
+import { getTappProviderId, selectTappProviderById } from "../helpers/provider"
 
 export function ActiveTapplet() {
   const { t } = useTranslation("components")
